Finalize initial instrumentation
