+++ package.json
@@ -70,6 +70,7 @@
     "@babel/core": "^7.16.0",
     "@istanbuljs/schema": "^0.1.3",
     "@syntest/framework": "^0.1.1",
+    "@types/babel-core": "^6.25.7",
     "istanbul-lib-instrument": "^5.1.0"
   }
 }
